+++ internal/utilities.go
@@ -1,0 +1,19 @@
+package internal
+
+import (
+	"errors"
+	"reflect"
+)
+
+// CallJobFuncWithParams call a function with parameters
+func CallJobFuncWithParams(jobFunc interface{}, params []interface{}) ([]reflect.Value, error) {
+	f := reflect.ValueOf(jobFunc)
+	if len(params) != f.Type().NumIn() {
+		return nil, errors.New("length of inputs don't match with length of function input")
+	}
+	in := make([]reflect.Value, len(params))
+	for k, param := range params {
+		in[k] = reflect.ValueOf(param)
+	}
+	return f.Call(in), nil
+}