+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat(ll): linklist structure added
 
 .. 1.0.0 (yyyy-mm-dd)
 .. ------------------